chore(backend): abstract list routes
