+++ packages/backend/src/open_payments/payment_pointer/routes.ts
@@ -1,5 +1,11 @@
-import { PaymentPointerContext } from '../../app'
+import { PaymentPointerSubresource } from './model'
+import { PaymentPointerSubresourceService } from './service'
+import { PaymentPointerContext, ListContext } from '../../app'
 import { IAppConfig } from '../../config/app'
+import {
+  getPageInfo,
+  parsePaginationQueryParameters
+} from '../../shared/pagination'
 
 interface ServiceDependencies {
   config: IAppConfig
@@ -33,4 +39,40 @@ export async function getPaymentPointer(
     assetScale: ctx.paymentPointer.asset.scale,
     authServer: deps.config.authServerGrantUrl
   }
+}
+
+interface ListSubresourceOptions<M extends PaymentPointerSubresource> {
+  ctx: ListContext
+  getPaymentPointerPage?: PaymentPointerSubresourceService<M>['getPaymentPointerPage']
+  toBody: (model: M) => Record<string, unknown>
+}
+
+export const listSubresource = async <M extends PaymentPointerSubresource>({
+  ctx,
+  getPaymentPointerPage,
+  toBody
+}: ListSubresourceOptions<M>) => {
+  const pagination = parsePaginationQueryParameters(ctx.request.query)
+  const page = await getPaymentPointerPage({
+    paymentPointerId: ctx.paymentPointer.id,
+    pagination,
+    clientId: ctx.clientId
+  })
+  const pageInfo = await getPageInfo(
+    (pagination) =>
+      getPaymentPointerPage({
+        paymentPointerId: ctx.paymentPointer.id,
+        pagination,
+        clientId: ctx.clientId
+      }),
+    page
+  )
+  const result = {
+    pagination: pageInfo,
+    result: page.map((item: M) => {
+      item.paymentPointer = ctx.paymentPointer
+      return toBody(item)
+    })
+  }
+  ctx.body = result
 }
